refactor: add errorHandler in administradoresService

+++ backend/src/services/administradores.ts
@@ -14,35 +14,64 @@ export class AdministradorService {
     };
 
     async pegarUsuarioPorId(id: number): Promise<Usuario | null> {
-        return await prisma.usuario.findUnique({
+        const idPego = await prisma.usuario.findUnique({
             where: { id }
         });
-        
+        if(!idPego){
+            throw new AppError('Id de usuário não encontrado', 404)
+        }
+        return idPego;
     };
 
     async criarUsuario(usuarioData: Prisma.UsuarioCreateInput): Promise<Usuario> {
         const senhaHashada = await hash(usuarioData.senha, 10)
 
-        return await prisma.usuario.create({
+         if(usuarioData.senha.length < 4){
+            throw new AppError('A senha deve conter no mínimo 4 caracteres', 406)
+        }
 
+        const usuarioCriado = await prisma.usuario.create({
+ 
             data: {...usuarioData,
             senha: senhaHashada
             },
         });
+
+        if(!usuarioCriado){
+            throw new AppError('Não foi possível criar o usuário')
+        }
+            return usuarioCriado
     };
 
     async atualizarUsuario(id: number, usuarioData: Prisma.UsuarioUpdateInput): Promise<Usuario> {
-        return await prisma.usuario.update({
+        const usuarioExiste = await prisma.usuario.findUnique({
+            where: { id }
+        })
+        if(!usuarioExiste){
+            throw new AppError("Id de usuário não encontrado", 404)
+        }
+        
+        const atualizaUsuario = await prisma.usuario.update({
             where: { id },
             data: usuarioData
         });
-        
+
+        return atualizaUsuario
     };
     
     async deletarUsuario(id: number): Promise<Usuario> {
-        return await prisma.usuario.delete({
+        const usuarioExiste = await prisma.usuario.findUnique({
+            where: { id }
+        })
+        if(!usuarioExiste){
+            throw new AppError("Id de usuário não encontrado", 404)
+        }
+
+        const deletaUsuario = await prisma.usuario.delete({
             where: { id }
         });
+        
+        return deletaUsuario
     };
 }
